+++ modules/notebook/src/test/kotlin/org/kalasim/examples/er/SurrogateER.kt
@@ -1,0 +1,65 @@
+package org.kalasim.examples.er
+
+import org.jetbrains.kotlinx.dataframe.api.toDataFrame
+import org.jetbrains.kotlinx.dataframe.io.writeCSV
+import org.kalasim.analysis.InteractionEvent
+import org.kalasim.analysis.ResourceActivityEvent
+import org.kalasim.collect
+import org.kalasim.exponential
+import org.kalasim.misc.QuartoUtil
+import kotlin.io.path.Path
+import kotlin.time.Duration.Companion.days
+import kotlin.time.Duration.Companion.hours
+
+object BigER {
+    @JvmStatic
+    fun main(args: Array<String>) {
+        val outputDir = Path("modules/notebook/src/test/kotlin/org/kalasim/examples/er")
+
+        val er = EmergencyRoom(
+            numPhysicians = 50,
+            numRooms = 30,
+            patientArrival = exponential(0.05.hours),
+            enableInternalMetrics = false
+        )
+        er.apply {
+
+            val rae= collect<ResourceActivityEvent> (  )
+            val interactions= collect<InteractionEvent> (  )
+            run(365.days)
+//            SessionPrefs.RENDER_BACKEND = LocalR()
+
+//            println(er.waitingLine.statistics)
+////            er.waitingLine.sizeTimeline.display().show()
+//
+//            er.waitingLine.lengthOfStayStatistics.display().show()
+//            er.waitingLine.sizeTimeline.display().show()
+
+            if(false) {
+//            rae.toDataFrame().plot(xmin to 0.0)
+                rae.toDataFrame().writeCSV(outputDir.resolve("er.bookings.csv").toFile())
+                interactions.toDataFrame().writeCSV(outputDir.resolve("er.interactions.csv").toFile())
+
+                QuartoUtil.runQuartoInDir(outputDir, "er_analysis.R")
+            }
+        }
+
+    }
+}
+
+object ProfileER {
+    @JvmStatic
+    fun main(args: Array<String>) {
+        val er = EmergencyRoom(
+            numPhysicians = 5,
+            numRooms = 10,
+            patientArrival = exponential(0.05.hours),
+            enableInternalMetrics = false
+        )
+        er.apply {
+            run(10000.days)
+        }
+    }
+}
+
+